# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Document APIs",
        "contact": {}
    },
    "paths": {
...
        "/cms/v1/users": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Cms"
                ],
                "summary": "List Users",
                "parameters": [
                    {
                        "enum": [
                            "enable",
                            "disable"
                        ],
                        "type": "string",
                        "x-enum-varnames": [
                            "DataStatusEnable",
                            "DataStatusDisable"
                        ],
                        "name": "data_status",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 10,
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "role_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "created_at.desc",
                        "name": "sorts",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.UserCmsDto"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Cms"
                ],
                "summary": "Create User",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/db.UserCmsData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.UserCmsDto"
                        }
                    }
                }
            }
        },
        "/cms/v1/users/{user_id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Cms"
                ],
                "summary": "Update User",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user_id",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/db.UserCmsData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.UserCmsDto"
                        }
                    }
                }
            }
        },
        "/cms/v1/users/{user_id}/reset-password": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Cms"
                ],
                "summary": "Reset Password For User",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user_id",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.UserCmsDto"
                        }
                    }
                }
            }
        },
...
        "/webhook/v1": {
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "tags": [
                    "Webhook"
                ],
                "summary": "Webhook",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/db.M"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.M"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "db.M": {
            "type": "object",
            "additionalProperties": {}
        },
...
        "db.UserCmsData": {
            "type": "object",
            "required": [
                "data_status",
                "email",
                "name",
                "phone",
                "role_ids",
                "username"
            ],
            "properties": {
                "data_status": {
                    "$ref": "#/definitions/enum.DataStatus"
                },
                "email": {
                    "type": "string",
                    "example": "aloha@email.com"
                },
                "name": {
                    "type": "string",
                    "example": "Aloha"
                },
                "phone": {
                    "type": "string",
                    "example": "0973123456"
                },
                "role_ids": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "671db9eca1f1b1bdbf3d4617"
                    ]
                },
                "username": {
                    "type": "string",
                    "example": "aloha"
                }
            }
        },
        "db.UserCmsDto": {
            "type": "object",
            "properties": {
                "data_status": {
                    "$ref": "#/definitions/enum.DataStatus"
                },
                "email": {
                    "type": "string",
                    "example": "aloha@email.com"
                },
                "name": {
                    "type": "string",
                    "example": "Aloha"
                },
                "phone": {
                    "type": "string",
                    "example": "0973123456"
                },
                "role_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "671db9eca1f1b1bdbf3d4617"
                    ]
                },
                "updated_at": {
                    "type": "string",
                    "example": "2006-01-02T03:04:05Z"
                },
                "updated_by": {
                    "type": "string",
                    "example": "editor"
                },
                "user_id": {
                    "type": "string",
                    "example": "671db9eca1f1b1bdbf3d4618"
                },
                "username": {
                    "type": "string",
                    "example": "aloha"
                }
            }
        },
...
        "enum.DataStatus": {
            "type": "string",
            "enum": [
                "enable",
                "disable"
            ],
            "x-enum-varnames": [
                "DataStatusEnable",
                "DataStatusDisable"
            ]
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        },
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "Header"
        }
    }
}